use integers instead of numerics
for municipality IDs after conversion from column names

+++ R/swcGetMapping.R
@@ -286,8 +286,8 @@ getMunicipalityMappingWorker <- function(swc, hist.list.from, mid.from, hist.lis
   )
 
   ff <- Matrix::summary(f)
-  ff$from <- as.numeric(rownames(f)[ff$i])
-  ff$to <- as.numeric(colnames(f)[ff$j])
+  ff$from <- as.integer(rownames(f)[ff$i])
+  ff$to <- as.integer(colnames(f)[ff$j])
   ff <- plyr::arrange(ff, get("from"))
   ff$i <- NULL
   ff$j <- NULL
